Move board struct creation to Board module
